+++ projects/ui/src/lib/components/po-dynamic/po-dynamic-form/po-dynamic-form-fields/po-dynamic-form-fields.component.ts
@@ -1,34 +1,10 @@
 import { Component, OnChanges, QueryList, SimpleChanges, ViewChildren } from '@angular/core';
 import { ControlContainer, NgForm } from '@angular/forms';
-import { HttpClient } from '@angular/common/http';
 import { TitleCasePipe } from '@angular/common';
 
-import { mapObjectByProperties } from '../../../../utils/util';
-
 import { PoDynamicFieldValidation } from './po-dynamic-form-field-validation.interface';
-import { PoDynamicFormField } from '../po-dynamic-form-field.interface';
 import { PoDynamicFormFieldsBaseComponent } from './po-dynamic-form-fields-base.component';
-
-const fieldProperties = [
-  'columns',
-  'required',
-  'options',
-  'optionsMulti',
-  'optionsService',
-  'searchService',
-  'mask',
-  'pattern',
-  'minLength',
-  'maxLength',
-  'disabled',
-  'help',
-  'booleanTrue',
-  'booleanFalse',
-  'maxValue',
-  'minValue',
-  'rows',
-  'secret'
-];
+import { PoDynamicFormValidationService } from './po-dynamic-form-fields-validation.service';
 
 /**
  * @docsPrivate
@@ -40,13 +16,14 @@ const fieldProperties = [
 @Component({
   selector: 'po-dynamic-form-fields',
   templateUrl: 'po-dynamic-form-fields.component.html',
-  viewProviders: [ { provide: ControlContainer, useExisting: NgForm } ]
+  viewProviders: [ { provide: ControlContainer, useExisting: NgForm } ],
+  providers: [ PoDynamicFormValidationService ]
 })
 export class PoDynamicFormFieldsComponent extends PoDynamicFormFieldsBaseComponent implements OnChanges {
 
   @ViewChildren('component') component: QueryList<{ focus: () => void }>;
 
-  constructor(titleCasePipe: TitleCasePipe, private http: HttpClient) {
+  constructor(titleCasePipe: TitleCasePipe, private validationService: PoDynamicFormValidationService) {
     super(titleCasePipe);
   }
 
@@ -56,7 +33,7 @@ export class PoDynamicFormFieldsComponent extends PoDynamicFormFieldsBaseCompone
     }
   }
 
-  onChangeFieldValue(index) {
+  onChangeFieldValue(index: number) {
     this.validateField(index);
     this.validateFields(index);
   }
@@ -67,15 +44,10 @@ export class PoDynamicFormFieldsComponent extends PoDynamicFormFieldsBaseCompone
       value: this.value,
     };
 
-    if (typeof this.validate === 'string') {
-      this.http.post(this.validate, changeValue).subscribe((response: any) => {
-        this.applyValidateFields(response);
-      });
-    } else {
-      const validatedFields = this.validate(changeValue);
-      this.applyValidateFields(validatedFields);
-    }
+    const validatedFields = await this.validationService.validateFields(this.validate, changeValue);
 
+    // TO DO: disabled form
+    this.applyValidateFields(validatedFields);
   }
 
   private applyValidateFields(validatedFields: any) {
@@ -94,16 +66,14 @@ export class PoDynamicFormFieldsComponent extends PoDynamicFormFieldsBaseCompone
 
   async validateField(index: number) {
     const field = this.fields[index];
-    const changedValue = { value: this.value[field.property], field: field.property };
+    const value = this.value[this.fields[index].property];
 
     const previousDisabled = field.disabled;
     this.visibleFields[index].disabled = true;
 
     try {
-      const fieldValidation: PoDynamicFieldValidation = typeof field.validate === 'string' ?
-        await this.validateFieldOnServer(field.validate, changedValue) : field.validate(changedValue);
-
-      this.applyValidationField(index, fieldValidation);
+      const validatedField = await this.validationService.validateField(field, value);
+      this.applyValidationField(index, validatedField);
 
     } catch {
       this.visibleFields[index].disabled = previousDisabled;
@@ -111,9 +81,7 @@ export class PoDynamicFormFieldsComponent extends PoDynamicFormFieldsBaseCompone
   }
 
   private applyValidationField(index: number, validatedField: PoDynamicFieldValidation) {
-    const validatedFieldClean = mapObjectByProperties(validatedField.field, fieldProperties, true);
-
-    this.fields[index] = { ...this.fields[index], ...validatedFieldClean };
+    this.fields[index] = { ...this.fields[index], ...validatedField };
     this.fields = [...this.fields];
     this.value[this.fields[index].property] = validatedField.value;
 
@@ -121,10 +89,6 @@ export class PoDynamicFormFieldsComponent extends PoDynamicFormFieldsBaseCompone
       this.component.toArray()[index].focus();
     }
 
-  }
-
-  private validateFieldOnServer(url: string, changedValue: { value: any; field: string; }) {
-    return this.http.post(url, changedValue).toPromise();
   }
 
   trackBy(index) {
